+++ am_list/runtime/queries/go/all_functions.scm
@@ -3,3 +3,21 @@
 
  (function_declaration
    name: (identifier) @func.name))
+
+((package_clause
+   (package_identifier) @pack.name)
+
+ (method_declaration
+   receiver: (parameter_list
+               (parameter_declaration
+                          type: (type_identifier) @type.name))
+   name: (field_identifier) @func.name))
+
+((package_clause
+   (package_identifier) @pack.name)
+
+ (method_declaration
+   receiver: (parameter_list
+               (parameter_declaration
+                          type: (pointer_type (type_identifier) @type.name)))
+   name: (field_identifier) @func.name))